Replace the yield with an isb on Arm.
The yield instruction is treated as a nop on Arm processors which is very
different than the x86 pause instruction that stalls execution for ~40 cycles.

An ISB serializes the pipeline and has been shown to be roughly analogous
to the pause delays and is used is other databases for spinloops and adaptive
spin loops where not hammering the cache line is important.